requirements.txt
streamlit
streamlit-folium
folium
pandas
requests
python-docx
